Move ares data manipulation to container level. Update expected proptypes
