+++ src/app/containers/Video/index.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import Figure from '@bbc/psammead-figure';
+import deepGet from '../../helpers/json/deepGet';
 import Video from '../../components/Video';
 import Caption from '../Caption';
 
@@ -16,9 +17,32 @@ const VideoContainer = ({ blocks }) => {
     return null;
   }
 
+  const nestedModel = deepGet(['model', 'blocks', 0, 'model'], aresMediaBlock);
+  const kind = deepGet(['subType'], nestedModel);
+  const pid = deepGet(['id'], nestedModel);
+  const title = deepGet(['title'], nestedModel);
+  const version = deepGet(['versions', 0], nestedModel);
+  const duration = deepGet(['duration'], version);
+  const versionID = deepGet(['versionId'], version);
+  const holdingImageUrl = deepGet(
+    ['blocks', 1, 'model', 'blocks', 0, 'model', 'locator'],
+    aresMediaBlock.model,
+  );
+  const items = {
+    versionID,
+    kind,
+    duration,
+  };
+
   return (
     <Figure>
-      <Video {...aresMediaBlock} />
+      <Video
+        pid={pid}
+        kind={kind}
+        title={title}
+        items={items}
+        holdingImageUrl={holdingImageUrl}
+      />
       {captionBlock ? <Caption block={captionBlock} video /> : null}
     </Figure>
   );
