+++ src/app/models/propTypes/index.js
@@ -1,4 +1,4 @@
-import { arrayOf, bool, shape, string } from 'prop-types';
+import { arrayOf, bool, number, shape, string } from 'prop-types';
 import { textBlockPropTypes } from './text';
 import { imageBlockPropTypes } from './image';
 
@@ -41,12 +41,15 @@ export const videoPropTypes = {
 };
 
 export const videoComponentPropTypes = {
-  duration: string,
-  imageLocator: string,
-  kind: string,
-  rawImageSrc: string,
-  versionID: string,
-  videoLocator: string,
+  pid: string.isRequired,
+  kind: string.isRequired,
+  title: string.isRequired,
+  items: shape({
+    versionID: string.isRequired,
+    kind: string.isRequired,
+    duration: number.isRequired,
+  }).isRequired,
+  holdingImageUrl: string.isRequired,
 };
 
 const baseDefaultPropTypes = {
